requirements.txt
absl-py==0.9.0
aiofiles==0.5.0
aiohttp==3.6.2
APScheduler==3.6.3
astor==0.8.1
async-generator==1.10
async-timeout==3.0.1
attrs==19.3.0
Automat==20.2.0
awscli==1.18.124
boto3==1.14.47
botocore==1.17.47
bottle==0.12.18
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
click==7.1.2
cloudpickle==1.3.0
colorama==0.4.3
colorclass==2.2.0
coloredlogs==10.0
colorhash==1.0.2
ConfigArgParse==0.13.0
constantly==15.1.0
cryptography==3.0
cycler==0.10.0
decorator==4.4.2
dnspython==1.16.0
docopt==0.6.2
docutils==0.15.2
fakeredis==0.10.3
fbmessenger==6.0.0
Flask==1.1.2
Flask-Cors==3.0.8
Flask-JWT-Simple==0.0.3
future==0.17.1
gast==0.2.2
gevent==1.5.0
google-auth==1.20.1
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
greenlet==0.4.16
grpcio==1.31.0
h11==0.8.1
h2==3.2.0
h5py==2.10.0
hpack==3.0.0
hstspreload==2020.8.18
httplib2==0.18.1
httptools==0.1.1
httpx==0.9.3
humanfriendly==8.2
hyperframe==5.2.0
hyperlink==20.0.1
idna==2.10
importlib-metadata==1.7.0
incremental==17.5.0
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.10.0
joblib==0.16.0
jsonpickle==1.4.1
jsonschema==3.2.0
kafka-python==1.4.7
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.2.0
klein==17.10.0
Markdown==3.2.2
MarkupSafe==1.1.1
matplotlib==3.2.2
mattermostwrapper==2.2
mock==4.0.2
multidict==4.7.6
networkx==2.4
numpy==1.19.1
oauth2client==4.1.3
oauthlib==3.1.0
opt-einsum==3.3.0
packaging==20.4
pika==1.1.0
prompt-toolkit==2.0.10
protobuf==3.13.0
psycopg2-binary==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydot==1.4.1
PyHamcrest==2.0.2
PyJWT==1.7.1
pykwalify==1.7.0
pymongo==3.8.0
pyparsing==2.4.7
pyrsistent==0.16.0
PySocks==1.7.1
python-crfsuite==0.9.7
python-dateutil==2.8.1
python-engineio==3.12.1
python-socketio==4.5.1
python-telegram-bot==12.8
pytz==2019.3
PyYAML==5.3.1
questionary==1.5.2
rasa==1.10.11
rasa-core==0.14.5
rasa-core-sdk==0.14.0
rasa-nlu==0.15.1
rasa-sdk==1.10.2
redis==3.5.3
regex==2020.6.8
requests==2.24.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
rfc3986==1.4.0
rocketchat-API==1.3.1
rsa==4.5
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
s3transfer==0.3.3
sanic==19.12.2
Sanic-Cors==0.10.0.post3
sanic-jwt==1.4.1
Sanic-Plugins-Framework==0.9.3
scikit-learn==0.22.2.post1
scipy==1.5.2
simplejson==3.17.2
six==1.15.0
sklearn-crfsuite==0.3.6
slackclient==2.8.0
sniffio==1.1.0
SQLAlchemy==1.3.19
tabulate==0.8.7
tensorboard==2.1.1
tensorflow==2.1.1
tensorflow-addons==0.7.1
tensorflow-estimator==2.1.0
tensorflow-hub==0.8.0
tensorflow-probability==0.9.0
termcolor==1.1.0
terminaltables==3.1.0
tornado==6.0.4
tqdm==4.45.0
twilio==6.26.3
Twisted==20.3.0
typing==3.7.4.3
typing-extensions==3.7.4.2
tzlocal==2.1
ujson==2.0.3
urllib3==1.25.10
uvloop==0.14.0
wcwidth==0.2.5
webexteamssdk==1.3
websocket-client==0.54.0
websockets==8.1
Werkzeug==1.0.1
wrapt==1.12.1
yarl==1.5.1
zipp==3.1.0
zope.interface==5.1.0